From: Jonathan Stowe <gellyfish@gellyfish.com>
This is the residue of the warning fixes I have made and which haven't
been applied before I start a new working copy :)


git-svn-id: https://svn.parrot.org/parrot/trunk@1315 d31e2699-5ff4-0310-a27c-f18f2fbe73fe

+++ packfile.c
@@ -125,8 +125,8 @@ PackFile_check_segment_size(opcode_t segment_size, const char *debug)
 
     if (segment_size % sizeof(opcode_t)) {
         fprintf(stderr,
-                "PackFile_unpack: Illegal %s table segment size %ld (must be multiple of %i)!\n",
-                debug, segment_size, sizeof(opcode_t));
+                "PackFile_unpack: Illegal %s table segment size %ld (must be multiple of %ld)!\n",
+                debug, segment_size, (long)sizeof(opcode_t));
         return 0;
     }
     return 1;
